% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{compose}
\title{Compose a new Rmarkdown document based on a template}
\usage{
compose(file, template, package = NULL, create_dir = "default",
  edit = TRUE, ...)
}
\arguments{
\item{file}{File name for the draft}

\item{template}{Template to use as the basis for the draft. This is either
the full path to a template directory or the name of a template directory
within the \code{rmarkdown/templates} directory of a package.}

\item{package}{(Optional) Name of package where the template is located.}

\item{create_dir}{\code{TRUE} to create a new directory for the document
(the "default" setting leaves this behavior up to the creator of the
template).}

\item{edit}{\code{TRUE} to edit the template immediately}

\item{...}{additional metadata variables to include in the YAML front matter.
NOTE: dots within variable names will be replaced with hyphens in the
output (e.g., \code{font.family} will become \code{font-family}).}
}
\description{
Compose a new Rmarkdown document based on a template
}
\examples{
compose("report.Rmd", author = "Aaron Wolen")

}
