% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod6_compose.R
\name{f_aggregation}
\alias{f_aggregation}
\title{Outputs for aggregation sub-tab}
\usage{
f_aggregation(
  coin,
  dset = NULL,
  aggregation_type,
  weight_set = NULL,
  min_dat_avail = NULL
)
}
\arguments{
\item{coin}{The coin}

\item{dset}{Data set to operate on. Leave \code{NULL} unless we allow to change operation order.}

\item{aggregation_type}{One of \code{"a_amean"} (arithmetic mean) or \code{"a_gmean"} (geometric mean).}

\item{weight_set}{A string: any one of the entries returned by \code{get_weight_sets_names()}.}

\item{min_dat_avail}{A number between 0 and 1.}
}
\value{
A list with \code{.$coin} the coin and \code{.$df_aggregated} the data frame of results to display.
}
\description{
This takes a coin and aggregates the selected data set. It returns the updated
coin plus a data frame of results that is sorted from highest to lowest-scoring
units.
}
\details{
By default this will use the "Normalised" data set and in fact if this dset is
not present it will show a warning. This is because aggregating indicators that are
not normalised is usually a bad idea - they may be on completely different scales.
We could make normalisation a hard constraint in the app. To discuss.

As it is this function will only allow one aggregation type for all levels. Although COINr can
take a different aggregation type for each level, we could move to that later if needed. The
\code{aggregation_type} should be a dropdown with two entries (\code{"a_mean"} and \code{"g_mean"}).

\code{weight_set} should be a dropdown which consists of the entries returned by \code{get_weight_sets_names()}.
This will return a character vector of names of any weight sets in the coin, plus "Equal". Weight
sets can be generated by the user in a later tab.

The \code{min_dat_avail} argument should be a slider input between 0 and 1, with default 1. This controls
the proportion of data that is needed \emph{within each aggregation group} to calculate an aggregate score.
See the \href{https://bluefoxr.github.io/COINr/articles/aggregate.html#data-availability-limits}{aggregate vignette in COINr for more info}.

On the same tab there should also be a bar chart output and possibly a map, although I will make
separate functions for these.
}
