% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod3_imputation.R
\name{f_imputation}
\alias{f_imputation}
\title{Impute}
\usage{
f_imputation(coin, dset = NULL, imputation_method, imputation_group = NULL)
}
\arguments{
\item{coin}{A coin}

\item{dset}{The data set to operate on. If not defined, used a set order as defined by \code{f_select_dset()}.}

\item{imputation_method}{One of \code{"i_mean"}, \code{"i_median"}, \code{"i_mean_grp"}, \code{"i_median_grp"}.}

\item{imputation_group}{Only specify if \code{imputation_method \%in\% c("i_mean_grp", "i_median_grp")}, else \code{NULL}.
If specified, points to a group variable which was entered with the input data to use in group imputation.}
}
\value{
The updated coin
}
\description{
Takes a coin and imputation specifications and outputs coin with imputed data set.
}
\details{
If imputation method is set as \code{"i_mean_grp"} or \code{"i_median_grp"}, a group is required to be specified via
the \code{imputation_group} argument. These groups are any that were input with the input data and metadata. If no
groups were input, these methods will not work.

We could possibly add a missing data visualisation as well at some point.
}
