% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod5_normalisation.R
\name{f_normalisation}
\alias{f_normalisation}
\title{Normalise}
\usage{
f_normalisation(coin, dset = NULL, norm_method, norm_params = NULL)
}
\arguments{
\item{coin}{The coin}

\item{dset}{The data set to apply the operation to. Probably not used so leave \code{NULL} unless allow the user to change order.}

\item{norm_method}{A string: one of \code{"n_minmax"}, \code{"n_zscore"}, \code{"n_borda"}, \code{"n_prank"}, which defines
which normalisation method to use.}

\item{norm_params}{Additional parameters to support the normalisation method:
\itemize{
\item if \code{norm_method = "n_minmax"}, then \code{norm_params} should be a numerical vector \code{c(x_lower, x_upper)} where \code{x_lower} and \code{x_upper} are
the lower and upper bounds of the interval to scale indicators on. E.g. \code{c(0, 100)} scales to the 0-100 interval.
\item if \code{norm_method = "n_zscore"} then \code{norm_params} should be a numerical vector \code{c(x_mean, x_sd)} defining the mean and standard deviation
to use in rescaling. E.g. \code{c(0, 1)} scales all indicators to have mean 0 and standard deviation 1.
}

Otherwise leave \code{NULL}.}
}
\value{
The updated coin.
}
\description{
Takes the normalisation method and parameters as inputs, and returns the coin with normalised data set.
}
\details{
We will need some kind of conditional inputs in the app. For \code{"n_borda"} and \code{"n_prank"} methods, no additional parameters are needed,
but for \code{"n_minmax"} we would need two inputs for upper and lower bounds respectively. With the constraint that upper must be less than
lower. For \code{"n_zscore"} we would need again two inputs but labelled "mean" and "standard deviation" respectively. Standard deviation
constrained to be greater than zero.

This sub-tab ideally also has a scatter plot and this is now provided by the function
\code{f_plot_scatter()} to avoid regenerating the data treatment if the plot changes and vice versa.
}
