% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod4_treat.R
\name{f_outlier_treatment}
\alias{f_outlier_treatment}
\title{Treat data set}
\usage{
f_outlier_treatment(
  coin,
  dset = NULL,
  skew_thresh = 2,
  kurt_thresh = 3.5,
  max_winsorise = 5
)
}
\arguments{
\item{coin}{A coin}

\item{dset}{Data set to use}

\item{skew_thresh}{Absolute skew threshold}

\item{kurt_thresh}{Kurtosis threshold}

\item{max_winsorise}{Max number of points to Winsorise for each indicator.}
}
\value{
The updated coin
}
\description{
The expected inputs are skew and kurtosis thresholds (numbers > 0), and the maximum number of
points to Winsorise (integer >= 0). More flexibility could be added to this function, e.g.
disabling the log transformation or Winsorisation.
}
