% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod8_correlations.R
\name{f_plot_correlation}
\alias{f_plot_correlation}
\title{Correlation plots}
\usage{
f_plot_correlation(
  coin,
  dset = "Raw",
  group1,
  group2,
  level1 = 1,
  level2 = 1,
  only_with_parents = FALSE,
  family_plot = FALSE,
  show_values = FALSE,
  discrete_colours = TRUE,
  plot_by_groups = TRUE
)
}
\arguments{
\item{coin}{The coin}

\item{dset}{Data set to use, one of \code{get_dset_names()}}

\item{group1}{First group of indicators. See details.}

\item{group2}{Second group of indicators.}

\item{level1}{Level to take first group from. See details.}

\item{level2}{Level to take second group from. See details.}

\item{only_with_parents}{Logical (switch in app?)}

\item{family_plot}{Logical (switch in app?)}

\item{show_values}{Logical (switch in app?)}

\item{discrete_colours}{Logical (switch in app?)}

\item{plot_by_groups}{Logical (switch in app?)}
}
\value{
Plot generated by ggplot2
}
\description{
Plots the correlations between specified groups of indicators/aggregates according to specifications.
}
\details{
This calls \code{\link[COINr:plot_corr]{COINr::plot_corr()}} which is a quite flexible function but comes with some complexity. The
question is how much flexibility/complexity to pass to the user. At the moment it is set up expecting
the user inputs as described in the function arguments. I expect that we may need some adjustments when
this is hooked up to the UI, as it becomes clearer what is useful and what not.

Another question: if we want interactive plots we can feed this into \code{ggplotly()}. It will need a little
tweaking.

For the \code{group1} and \code{group2} args, the input to the \emph{function} should be one of
\code{get_indicator_codes(coin, with_levels = FALSE)}, whereas the \emph{dropdown} in the app should display instead
\code{get_indicator_codes(coin, with_levels = TRUE)}. The latter is hopefully clearer to display.

NOTE that for the \code{level1} an \code{level2} user inputs, these must be set up as follows:
\itemize{
\item if \code{dset == "Aggregated"} the dropdown for both should be \code{1 : get_n_levels(coin)}
\item else for any other \code{dset} the dropdowns should be hidden or should only contain a single entry: 1.
}

At the moment the function will set levels to 1 and issue a warning if the latter case is not followed.
}
