% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod5_normalisation.R
\name{f_plot_scatter_norm}
\alias{f_plot_scatter_norm}
\title{Plot scatter plot before/after normalisation}
\usage{
f_plot_scatter_norm(coin, df_normalised, icol_selected, dset1)
}
\arguments{
\item{coin}{The coin}

\item{df_normalised}{is the normalised data frame}

\item{icol_selected}{is the column index selected (to plot)}

\item{dset1}{The data set to plot on x axis}
}
\description{
Plot of selected indicator
}
