% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod3_imputation.R
\name{f_print_pre_imputation}
\alias{f_print_pre_imputation}
\title{Summarise missing data points before imputation}
\usage{
f_print_pre_imputation(l)
}
\arguments{
\item{l}{The list \code{r_coin_imputed()} in the server}
}
\description{
Summarise missing data points before imputation
}
