% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod1-overview.R
\name{f_read_input}
\alias{f_read_input}
\title{Read file inputs}
\usage{
f_read_input(path1, path2, input_type)
}
\arguments{
\item{path1}{First file path. If \code{input_type = "xlsx"} this will be the path to the
xlsx file. Otherwise, it will be the path to the first csv file which should contain
a data frame with indicator data in the \code{iData} format.}

\item{path2}{Second file path. If \code{input_type = "csv"} this should be the path to
a csv file with an \code{iMeta} table inside. If \code{input_type = "xlsx"} this argument is ignored.}

\item{input_type}{One of \code{"xlsx"} or \code{"csv"}.}
}
\value{
A coin-class object
}
\description{
Reads file inputs as either two csv files or one xlsx file, and returns a coin.
We can decide whether to allow the user to input csvs or xlsx, or just constrain them to use one.
For the moment I put both here.
}
\details{
If the input is csv, the two files must follow the formats specified in \url{https://bluefoxr.github.io/COINr/reference/new_coin.html}.
Otherwise if the input is xlsx, the same format is followed but both tables are in a single xlsx file
with tabs called "iData" and "iMeta".

Note that if there is a problem with the input files, this function will output error messages
which we should display somewhere, e.g. in a text output box, as they can help to point to
the problem and the solution for the user.
}
\examples{
# excel
# coin <- f_read_input("./data/example_input.xlsx", input_type = "xlsx")
# csv
# coin <- f_read_input("./data/iData_example.csv",
#                      "./data/iMeta_example.csv",
#                      input_type = "csv")
}
