% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod13_reweighting.R
\name{f_reweight_coin}
\alias{f_reweight_coin}
\title{Regenerate results}
\usage{
f_reweight_coin(coin, w)
}
\arguments{
\item{coin}{The coin}

\item{w}{can either be a named list with names as iCodes and values
as the new weights, OR as a data frame with columns "iCode" and "Weight" with
codes and corresponding weights.}
}
\value{
Updated coin
}
\description{
Regenerates results based on specified weights
}
