% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod15_sensitivity.R
\name{f_run_SA}
\alias{f_run_SA}
\title{Run sensitivity analysis}
\usage{
f_run_SA(
  coin,
  test_imputation,
  test_treat,
  test_aggregation,
  test_weights,
  perturb_weights,
  N_rep,
  run_UA
)
}
\arguments{
\item{coin}{The coin}

\item{test_imputation}{Logical: if TRUE tests the effect of switching imputation
on and off.}

\item{test_treat}{Logical: if TRUE tests the effect of switching outlier treatment
on and off.}

\item{test_aggregation}{Logical: if TRUE tests the effect of switching between
arithmetic and geometric mean.}

\item{test_weights}{Logical: if TRUE tests the effect of perturbing weights}

\item{perturb_weights}{Percentage perturbation to apply to weights, if \code{test_weights = TRUE}}

\item{N_rep}{The number of replications to run (true number is N_rep*(n_assumptions + 2))}

\item{run_UA}{Logical: if \code{TRUE} runs an uncertainty analysis rather than a sensitivity analysis.
This is intended to be set \code{TRUE} when only one uncertainty is investigated.}
}
\value{
List of sensitivity analysis results
}
\description{
For the moment, let's assume that users can only test sensitivity to operations
that they have actually run in the app. E.g. the user can only test the
imputation option if they have actually run the imputation tab. Otherwise, this
option will be greyed out in the app. Hence this function assumes that if a
switch is TRUE then that option has been run, else will return an error.
}
