% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{f_select_dset}
\alias{f_select_dset}
\title{Select data set to operate on}
\usage{
f_select_dset(coin, calling_function)
}
\arguments{
\item{coin}{A coin}

\item{calling_function}{The calling function, as a string.}
}
\value{
A data set name, as a string
}
\description{
Helper function which, given a calling function, returns the data set to
operate on. This is assuming the order of data functions as given in the
function definition.
}
