% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod7_stats.R
\name{f_stat_table}
\alias{f_stat_table}
\title{Table of stats and flags for indicators}
\usage{
f_stat_table(
  coin,
  dset = "Raw",
  dat_avail_thresh = 0.66,
  skew_thresh = 2,
  kurt_thresh = 3.5,
  duplicate_thresh = 0.5,
  collinear_thresh = 0.9,
  neg_corr_thresh = -0.4
)
}
\arguments{
\item{coin}{The coin}

\item{dset}{Name of data set to analyse.}

\item{dat_avail_thresh}{Number between 0 and 1.}

\item{skew_thresh}{Positive number. Probably we don't need to expose the user to this.}

\item{kurt_thresh}{Positive number. Probably we don't need to expose the user to this.}

\item{duplicate_thresh}{Number between 0 and 1: threshold for the max fraction of repeated
values. Above this will be flagged.}

\item{collinear_thresh}{Number between 0 and 1: correlation pairs above this will be flagged
as collinear.}

\item{neg_corr_thresh}{As previous but for negative correlations.}
}
\description{
Outputs a list with two data frames: \code{.$FlaggedStats} and \code{.$Flags}. The first
is a table to display which contains things like max, min, skew and kurt,
data availability and so on. The second is an equivalent-sized df but with
logical columns: this specifies which elements of the first df represent "flags".
E.g. if data availability for an indicator is below \code{dat_avail_thresh}, it will
return \code{TRUE}. The output of this function is intended to be fed into \code{f_style_screening_table()}
or something similar.
}
\details{
NOTE: we will have to think what the default for \code{dset} is. Normally it makes most
sense to look at the raw data, but at the moment this table is positioned in the
"results" end of the app. To discuss.
}
