% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod3_imputation.R
\name{f_style_imputation_table}
\alias{f_style_imputation_table}
\title{DT table with highlighted cells}
\usage{
f_style_imputation_table(
  df_display,
  df_highlight,
  highlight_background = "#ffc266",
  highlight_text = "red"
)
}
\arguments{
\item{df_display}{The data frame to display (imputed data)}

\item{df_highlight}{A data frame, same size as \code{df_display}, with only logical
values. Any cell with \code{TRUE} will highlight the corresponding cell in \code{df_display}.}

\item{highlight_background}{Colour to highlight background of highlighted cells}

\item{highlight_text}{Highlight text colour}
}
\value{
DT table
}
\description{
Adds highlighted cells for imputed points, and summary row for number
of missing points per column (indicator).
}
