% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod2_screening.R
\name{f_style_screening_table}
\alias{f_style_screening_table}
\title{Generate screened DT table}
\usage{
f_style_screening_table(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
A DT table
}
\description{
Uses DT to generate a table to display, with rows of any screened units
coloured differently to show they have been removed. Colours maybe to be
adjusted.
}
