% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod7_stats.R
\name{f_style_stats_table}
\alias{f_style_stats_table}
\title{Styling for indicator stats table}
\usage{
f_style_stats_table(l, table_caption = NULL, highlight_colour = "#f6bdb4")
}
\arguments{
\item{l}{A list - the output of \code{\link[=f_stat_table]{f_stat_table()}}}

\item{table_caption}{Optional caption}

\item{highlight_colour}{Colour to highlight flagged cells}
}
\value{
DT table
}
\description{
Styling for indicator stats table
}
