% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod2_screening.R
\name{f_unit_screening}
\alias{f_unit_screening}
\title{Generate screened data set}
\usage{
f_unit_screening(
  coin,
  dset = NULL,
  screening_type,
  dat_thresh = NULL,
  zero_thresh = NULL
)
}
\arguments{
\item{coin}{The coin class object}

\item{dset}{The data set to apply screening to. Defaults to \code{"Raw"} and should be left at
this default unless we change the order of data operations.}

\item{screening_type}{One of \code{c("byNA", "byzeros", "byNAandzeros")}.}

\item{dat_thresh}{Number between 0 and 1, from "Min. data availability" input.}

\item{zero_thresh}{Number between 0 and 1 from "Min. non-zero" input.}
}
\value{
The updated coin
}
\description{
Generates screened data set on the coin.
}
\details{
Note if we keep the order of data operations fixed, this should run always on the
"Raw" data set, so the \code{dset} argument probably doesn't need to be specified.
}
