% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{get_iCodes}
\alias{get_iCodes}
\title{Get iCodes}
\usage{
get_iCodes(coin, Level = 1)
}
\arguments{
\item{coin}{A coin}

\item{Level}{Index level, as integer}
}
\value{
Vector of codes
}
\description{
Get iCodes at specified level in the coin
}
