% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{load_defaults}
\alias{load_defaults}
\title{Load user defined defaults for metadata values}
\usage{
load_defaults(paths = NULL, file = getOption("composer.composer_file"),
  template = NULL, package = NULL)
}
\arguments{
\item{paths}{paths containing YAML files (for testing purposes)}

\item{file}{YAML file containing metadata values}
}
\description{
Check the user's home directory and current working directory for a
\code{.composer} YAML file.
}
