% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_countries}
\alias{valid_countries}
\title{Valid country codes and names}
\format{
A data frame with 249 rows and 4 variables.
}
\source{
\url{https://github.com/vincentarelbundock/countrycode}
}
\usage{
valid_countries
}
\description{
This is a data frame of ISO3c codes and names of countries. It is used to
check the user input, to see whether the codes are ISO3 codes, and therefore
we can plot the data on the map. The codes and names were taken from the
"countrycode" package on 26/06 and may need to be updated every now and then.
Note that while ISO3 codes are a standard, country names can vary significantly.
}
\keyword{datasets}
