% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod1-overview.R
\name{validate_iData}
\alias{validate_iData}
\title{Verbose check on iData tab}
\usage{
validate_iData(iData)
}
\arguments{
\item{iData}{Table imported from user input spreadsheet.}
}
\value{
Message as text, or \code{NULL} if no errors.
}
\description{
Checks and messages to pass to the user for basic errors in the iData tab of
the input spreadsheet.
Note COINr does many of these checks anyway, but the intention is to generate
some user-friendly versions here. Called in \code{\link[=f_read_input]{f_read_input()}}.
}
