% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{weights_slider}
\alias{weights_slider}
\title{Generate 0-1 weights slider}
\usage{
weights_slider(inputId, initial_weight = 0.5, label = NULL, ns_id = NULL)
}
\arguments{
\item{inputId}{Id of slider}

\item{initial_weight}{The initial weight of the slider}

\item{label}{Display label of slider}

\item{ns_id}{Optional namespace ID}
}
\description{
A single 0 -1 slider for weights
}
